Query all address balances
Tendermint offers JSON:RPC endpoint that allows the submission of Cosmos
gRPC queries. This allows easy queries of the Mainnet through the public
endpoint.

Local testing is possible by running a test node with docker-compose.
This test endpoint can be used to run test queries against the gRPC
endpoint and the JSON::RPC endpoint.

Cosmos Query protobuf ruby generation

Note: gogoproto options with castrepeated need to be removed. This
option is not in the gogoproto definition.